+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 2.6.0 - 2017-01-31
+
+### Added
+- [#99](https://github.com/zendframework/zend-http/pull/99) added
+  TimeoutException for cURL adapter.
+- [#98](https://github.com/zendframework/zend-http/pull/98) added connection
+  timeout (`connecttimeout`) for cURL and Socket adapters.
+- [#97](https://github.com/zendframework/zend-http/pull/97) added support to
+  `sslcafile` and `sslcapath` to cURL adapter.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- Nothing.
+
 ## 2.5.6 - 2017-01-31
 
 ### Added
@@ -27,6 +49,8 @@ All notable changes to this project will be documented in this file, in reverse 
   Security Policy CSP HTTP header when it is `none` (empty value).
 - [#92](https://github.com/zendframework/zend-http/pull/92) fixes the flatten
   cookies value for array value (also multidimensional).
+- [#34](https://github.com/zendframework/zend-http/issues/33) fixes the
+  standard separator (&) for application/x-www-form-urlencoded.
 
 ## 2.5.5 - 2016-08-08